Add more rendering options
* Tighten up an off-by-one error with regard to even vs.
odd sample sizes
* Also added a Bell 103 sample generated with minimodem
since this serves as a known signal from which I can
expect a particular decoded set of bits, whereas I was
starting to have some doubts about my previous SAME sample
matching the SAME specification

+++ Core/AudioAnalysis/SamplingResult.cs
@@ -7,7 +7,7 @@ namespace Core.AudioAnalysis
         public IList<short> Samples { get; set; }
         public int SampleRate { get; set; }
         public int BitsPerSample { get; set; }
-        public int FileLengthInMilliseconds { get; set; }
+        public double FileLengthInMilliseconds { get; set; }
         public int FileLengthInBytes { get; set; }
     }
 }
